restructured/refactored trading blueprint
