+++ app/templates/trading/stock_page.html
@@ -10,9 +10,9 @@
         <ol class="breadcrumb">
           <li class="breadcrumb-item"><a href="{{ url_for('main.index') }}">Home</a></li>
             {% if current_user.is_authenticated %}
-            <li class="breadcrumb-item"><a href="{{ url_for('main.stocks') }}">Stocks</a></li>
+            <li class="breadcrumb-item"><a href="{{ url_for('trading.stocks') }}">Stocks</a></li>
             {% else %}
-            <li class="breadcrumb-item"><a href="{{ url_for('main.sample_stocks') }}">Stocks</a></li>  
+            <li class="breadcrumb-item"><a href="{{ url_for('trading.sample_stocks') }}">Stocks</a></li>  
             {% endif %}
           <li class="breadcrumb-item active" aria-current="page">{{ stock.symbol }}</li>
         </ol>
@@ -23,7 +23,7 @@
     <p>Select a trading strategy</p>
 
     <div class="list-group">
-      <a href="{{ url_for('main.stock_plot', symbol=stock.symbol, strategy='sma') }}" class="list-group-item list-group-item-action" aria-current="true">
+      <a href="{{ url_for('trading.stock_plot', symbol=stock.symbol, strategy='sma') }}" class="list-group-item list-group-item-action" aria-current="true">
         <div class="d-flex w-100 justify-content-between">
           <h5 class="mb-1">Simple Moving Averages (SMA)</h5>
         </div>
