+++ app/templates/trading/stocks.html
@@ -46,7 +46,7 @@
                 <td>{{ stock.symbol }}</td>
                 <td>{{ stock.name }}</td>
                 <td>
-                  <a class="btn btn-primary btn-sm" href="{{ url_for('main.stock_page', symbol=stock.symbol) }}" role="button">Go</a>
+                  <a class="btn btn-primary btn-sm" href="{{ url_for('trading.stock_page', symbol=stock.symbol) }}" role="button">Go</a>
                 </td>
             </tr>
            {% endfor %}
