+++ app/trading/__init__.py
@@ -1,6 +1,11 @@
+from flask import Blueprint
 from .strategy import (
     SMA,
     MeanReversion,
     OLS,
     Momentum
 )
+
+trading = Blueprint('trading', __name__)
+
+from . import views
